Value attribute in PQV,  REAL should be of type Double instead of BigDecimal to be inconsistent with ISO xsd. Following classes are modified: gov.nih.nci.iso21090.Real gov.nih.nci.iso21090.Pqv gov.nih.nci.iso21090.grid.dto.transform.iso.PQTransformer gov.nih.nci.iso21090.grid.dto.transform.iso.REALTransformer gov.nih.nci.iso21090.grid.dto.transform.iso.IVLPQTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.IVLREALTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.PQTransformerTest gov.nih.nci.iso21090.grid.dto.transform.iso.REALTransformerTest gov.nih.nci.iso21090.RealTest
SVN-Revision: 194

+++ localization/src/main/java/gov/nih/nci/iso21090/Real.java
@@ -82,8 +82,6 @@
  */
 package gov.nih.nci.iso21090;
 
-import java.math.BigDecimal;
-
 import org.apache.commons.lang.builder.EqualsBuilder;
 import org.apache.commons.lang.builder.HashCodeBuilder;
 
@@ -97,21 +95,21 @@ public final class Real extends Qty implements Cloneable {
 
     private static final long serialVersionUID = 1L;
 
-    private BigDecimal value;
+    private Double value;
 
     /**
      *
      * @return value
      */
-    public BigDecimal getValue() {
+    public Double getValue() {
         return value;
     }
 
     /**
      *
-     * @param value value of BigDecimal
+     * @param value value of Double
      */
-    public void setValue(BigDecimal value) {
+    public void setValue(Double value) {
         this.value = value;
     }
